refactor(SaveUseCase): refactor ConcurrentEventUserRepository to SaveUseCase
