+++ app/src/androidTest/java/com/android/unio/components/notification/NotificationTest.kt
@@ -14,7 +14,7 @@ import com.android.unio.model.event.EventViewModel
 import com.android.unio.model.image.ImageRepositoryFirebaseStorage
 import com.android.unio.model.notification.NotificationWorker
 import com.android.unio.model.notification.UnioNotification
-import com.android.unio.model.save.ConcurrentEventUserRepositoryFirestore
+import com.android.unio.model.usecase.SaveUseCaseFirestore
 import com.android.unio.model.search.SearchRepository
 import com.android.unio.model.search.SearchViewModel
 import com.android.unio.model.user.UserDeletionRepositoryFirestore
@@ -57,7 +57,7 @@ class NotificationTest : TearDown() {
   private lateinit var eventUserPictureRepositoryFirestore: EventUserPictureRepositoryFirestore
   @MockK
   private lateinit var concurrentEventUserRepositoryFirestore:
-      ConcurrentEventUserRepositoryFirestore
+          SaveUseCaseFirestore
   private lateinit var eventViewModel: EventViewModel
   private lateinit var searchViewModel: SearchViewModel
   private lateinit var context: Context
