+++ app/src/main/java/com/android/unio/model/event/EventViewModel.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import com.android.unio.model.association.AssociationRepository
 import com.android.unio.model.image.ImageRepository
-import com.android.unio.model.save.ConcurrentEventUserRepository
+import com.android.unio.model.usecase.SaveUseCase
 import com.android.unio.model.strings.StoragePathsStrings
 import com.android.unio.model.user.User
 import com.google.firebase.Firebase
@@ -33,7 +33,7 @@ constructor(
     private val imageRepository: ImageRepository,
     private val associationRepository: AssociationRepository,
     private val eventUserPictureRepository: EventUserPictureRepository,
-    private val concurrentEventUserRepository: ConcurrentEventUserRepository
+    private val saveUseCase: SaveUseCase
 ) : ViewModel() {
 
   /**
@@ -327,7 +327,7 @@ constructor(
       updatedUser.savedEvents.add(target.uid)
       Firebase.messaging.subscribeToTopic(target.uid)
     }
-    concurrentEventUserRepository.updateSave(
+    saveUseCase.updateSave(
         updatedUser,
         updatedEvent,
         {
