+++ app/src/test/java/com/android/unio/model/save/SaveUseCaseFirestoreTest.kt
@@ -3,6 +3,7 @@ package com.android.unio.model.save
 import com.android.unio.mocks.event.MockEvent
 import com.android.unio.mocks.user.MockUser
 import com.android.unio.model.event.EventRepository
+import com.android.unio.model.usecase.SaveUseCaseFirestore
 import com.android.unio.model.user.UserRepository
 import com.google.firebase.firestore.FirebaseFirestore
 import io.mockk.MockKAnnotations
@@ -11,14 +12,14 @@ import io.mockk.verify
 import org.junit.Before
 import org.junit.Test
 
-class ConcurrentEventUserRepositoryFirestoreTest {
+class SaveUseCaseFirestoreTest {
 
   @MockK private lateinit var db: FirebaseFirestore
   @MockK private lateinit var userRepository: UserRepository
   @MockK private lateinit var eventRepository: EventRepository
 
   private lateinit var concurrentEventUserRepositoryFirestore:
-      ConcurrentEventUserRepositoryFirestore
+          SaveUseCaseFirestore
 
   private val user = MockUser.createMockUser(uid = "1")
   private val event = MockEvent.createMockEvent(uid = "11")
@@ -28,7 +29,7 @@ class ConcurrentEventUserRepositoryFirestoreTest {
     MockKAnnotations.init(this, relaxed = true)
 
     concurrentEventUserRepositoryFirestore =
-        ConcurrentEventUserRepositoryFirestore(db, userRepository, eventRepository)
+        SaveUseCaseFirestore(db, userRepository, eventRepository)
   }
 
   @Test
